ajustes para mensagens e conserto do login sem ser user do twitter;
TODO:

- terminar visualização de grupos;
- criar visualização de categorias;
- permitir ver o timeline sem precisar mostrar as atualizacoes no painel (pcausa 

da OutOfMemoryException);
- atualizar ultimo status do usuario em seu nodo da rede;
- criar visualização de membros + ativos (vamos mesmo fazer esta?);
- implementar categorias para mostrar nas visualizações;
- consertar aba de favoritos (está repetindo informacoes);
- inserir no painel de atualizacoes a data e o link para ver na web;
- permitir varias buscas;
- classe Tweet: fazer imagem do usuario mostrar ele na rede;
- para painel de um unico usuario, nao mostrar a foto varias vezes;

- criar toolbar principal;
- criar menu configurações;
- fazer requisicoes com threads para nao trancar visualizacao;
- implementar menu bloquear usuario (mostrando um X no nodo);
- fazer algo com trends (#) - aba ou tag cloud;
- consertar ação de esconder amigos do nodo clicado;
- consertar cor de usuarios ligados por arestas invisiveis;
- adicionar filtro de nome na rede;
- parece que public timeline nao esta mais na pagina do twitter;
- implementar notificação quando novas atualizações/msgs chegarem;
- mostrar mensagens de erro e notificações na toolbar principal;
- permitir esconder/deletar nodos e inserir botao para ver todos;
- escolher forças de atração/repulsão melhores;
- setar tooltips do login;

+++ twitter4j/src/main/java/twitter4j/Twitter.java
@@ -1729,7 +1729,7 @@ public class Twitter implements java.io.Serializable {
      * @see <a href="http://apiwiki.twitter.com/Twitter-REST-API-Method%3A-direct_messages%C2%A0destroy">Twitter API Wiki / Twitter REST API Method: direct_messages destroy</a>
      * @since Twitter4J 2.0.1
      */
-    public DirectMessage destroyDirectMessage(int id) throws
+    public DirectMessage destroyDirectMessage(long id) throws
             TwitterException {
         return new DirectMessage(http.post(baseURL +
                 "direct_messages/destroy/" + id + ".xml", new PostParameter[0], true), this);
